fix: get all country test seperated
